+++ Explorer/Assets/DCL/Interaction/Raycast/Tests/ExecuteRaycastSystemShould.cs
@@ -1,5 +1,6 @@
 using Arch.Core;
 using CRDT;
+using CrdtEcsBridge.Components.ResetExtensions;
 using CrdtEcsBridge.Components.Special;
 using CrdtEcsBridge.ECSToCRDTWriter;
 using CrdtEcsBridge.Physics;
